refactor(app-route): share middleware header merge helper
