+++ packages/vinext/src/server/app-route-handler-response.ts
@@ -1,4 +1,5 @@
 import type { CachedRouteValue } from "../shims/cache.js";
+import { mergeMiddlewareResponseHeaders } from "./middleware-response-headers.js";
 
 export type RouteHandlerMiddlewareContext = {
   headers: Headers | null;
@@ -37,16 +38,7 @@ export function applyRouteHandlerMiddlewareContext(
   }
 
   const responseHeaders = new Headers(response.headers);
-  if (middlewareContext.headers) {
-    for (const [key, value] of middlewareContext.headers) {
-      const lowerKey = key.toLowerCase();
-      if (lowerKey === "set-cookie" || lowerKey === "vary") {
-        responseHeaders.append(key, value);
-      } else {
-        responseHeaders.set(key, value);
-      }
-    }
-  }
+  mergeMiddlewareResponseHeaders(responseHeaders, middlewareContext.headers);
 
   return new Response(response.body, {
     status: middlewareContext.status ?? response.status,
